chao bigint

+++ db/init.sql
@@ -1,7 +1,7 @@
 -- CREATE DATABASE NAME OWNER USER;
 
 CREATE TABLE IF NOT EXISTS farmacias (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     some_code VARCHAR(10) NOT NULL UNIQUE,
     name_farmacia TEXT NOT NULL,
     rif VARCHAR(255) NOT NULL UNIQUE,
@@ -10,9 +10,9 @@ CREATE TABLE IF NOT EXISTS farmacias (
 CREATE INDEX ix_farmacias_name ON farmacias (name_farmacia);
 
 CREATE TABLE IF NOT EXISTS caja_farmacia (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT NOT NULL,
+    id_farmacia INTEGER NOT NULL,
     nm_caja INTEGER NOT NULL,
     area VARCHAR(255) NOT NULL,
     CONSTRAINT fk_cajas_farmacia_id_farmacia FOREIGN KEY (id_farmacia) 
@@ -21,9 +21,9 @@ CREATE TABLE IF NOT EXISTS caja_farmacia (
 CREATE INDEX ix_caja_farmacia_id_farmacia ON caja_farmacia (id_farmacia);
 
 CREATE TABLE IF NOT EXISTS punto_venta (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT NOT NULL,
+    id_farmacia INTEGER NOT NULL,
     modelo VARCHAR(255) NOT NULL,
     banco TEXT NOT NULL,
     serial_code VARCHAR(255) NOT NULL UNIQUE,
@@ -35,9 +35,9 @@ CREATE INDEX ix_punto_venta_banco ON punto_venta (banco);
 CREATE INDEX ix_punto_venta_serial ON punto_venta (serial_code);
 
 CREATE TABLE IF NOT EXISTS equipo_pc (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     ip VARCHAR(255) NOT NULL,
     anydesk VARCHAR(255) DEFAULT '',
     sa_anydesk VARCHAR(255) DEFAULT NULL,
@@ -52,9 +52,9 @@ CREATE INDEX ix_equipo_pc_ip ON equipo_pc (ip);
 CREATE INDEX ix_equipo_pc_id_farmacia ON equipo_pc (id_farmacia);
 
 CREATE TABLE IF NOT EXISTS equipo_impresora (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     modelo_print TEXT NOT NULL,
     marca TEXT NOT NULL,
     area VARCHAR(255) NOT NULL,
@@ -65,11 +65,11 @@ CREATE TABLE IF NOT EXISTS equipo_impresora (
 CREATE INDEX ix_equipo_impresora_marca ON equipo_impresora (marca);
 
 CREATE TABLE IF NOT EXISTS cajas_asigne_equipo (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     resum_equipo TEXT DEFAULT NULL,
-    id_equipo BIGINT NOT NULL,
+    id_equipo INTEGER NOT NULL,
     observacion_asignacion TEXT DEFAULT NULL,
 
     CONSTRAINT fk_asigne_equipo_farmacia FOREIGN KEY (id_farmacia) 
@@ -83,10 +83,10 @@ CREATE INDEX ix_cajas_asigne_equipo_id_equipo ON cajas_asigne_equipo (id_equipo)
 CREATE UNIQUE INDEX ux_cajas_equipo_unico ON cajas_asigne_equipo (id_equipo);
 
 CREATE TABLE IF NOT EXISTS cajas_asigne_punto_venta (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
-    id_punto_venta BIGINT NOT NULL,
+    id_farmacia INTEGER,
+    id_punto_venta INTEGER NOT NULL,
     observacion_pos TEXT DEFAULT NULL,
 
     CONSTRAINT fk_asigne_pv_farmacia FOREIGN KEY (id_farmacia) 
@@ -99,9 +99,9 @@ CREATE UNIQUE INDEX ux_asigne_pv_equipo_unico ON cajas_asigne_punto_venta (id_pu
 CREATE INDEX ix_asigne_pv_id_punto_venta ON cajas_asigne_punto_venta (id_punto_venta);
 
 CREATE TABLE IF NOT EXISTS requerimientos_solicitudes (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     name_req VARCHAR(255) NOT NULL,
     descripcion TEXT DEFAULT 'sin Requerimiento',
     fix BOOL DEFAULT false,
@@ -118,10 +118,10 @@ CREATE INDEX ix_requerimientos_fecha_req ON requerimientos_solicitudes (fecha_re
 CREATE INDEX ix_requerimientos_name_req ON requerimientos_solicitudes (name_req);
 
 CREATE TABLE IF NOT EXISTS incidencias_pc (
-    id BIGSERIAL PRIMARY KEY,
-    id_farmacia BIGINT,
+    id SERIAL PRIMARY KEY,
+    id_farmacia INTEGER,
 	equipo_resum TEXT DEFAULT NULL,
-    id_equipo BIGINT NOT NULL,
+    id_equipo INTEGER NOT NULL,
     nombre_inci VARCHAR(255) NOT NULL,
     descripcion TEXT DEFAULT 'sin novedades',
     fecha_incidencia TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -138,9 +138,9 @@ CREATE INDEX ix_incidencias_pc_nombre ON incidencias_pc (nombre_inci);
 CREATE INDEX ix_incidencias_pc_fecha ON incidencias_pc (fecha_incidencia);
 
 CREATE TABLE IF NOT EXISTS incidencias_generales (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     nombre_inci VARCHAR(255) NOT NULL,
     descripcion TEXT DEFAULT 'sin novedades',
     fecha_registro TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -154,9 +154,9 @@ CREATE INDEX ix_incidencias_gen_nombre ON incidencias_generales (nombre_inci);
 CREATE INDEX ix_incidencias_gen_fecha ON incidencias_generales (fecha_registro);
 
 CREATE TABLE IF NOT EXISTS inventario_general (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
+    id_farmacia INTEGER,
     hardware TEXT NOT NULL,
     nota TEXT DEFAULT NULL,
     cantidad INTEGER DEFAULT 0,
@@ -169,13 +169,13 @@ CREATE INDEX ix_inventario_gen_hardware ON inventario_general (hardware);
 CREATE INDEX ix_inventario_farmacia_hardware ON inventario_general (id_farmacia, hardware);
 
 CREATE TABLE IF NOT EXISTS usuario_role (
-	id BIGSERIAL PRIMARY KEY,
+	id SERIAL PRIMARY KEY,
 	rolee VARCHAR (255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS usuario (
-    id BIGSERIAL PRIMARY KEY,
-    id_role BIGINT DEFAULT NULL,
+    id SERIAL PRIMARY KEY,
+    id_role INTEGER DEFAULT NULL,
     name_user VARCHAR(255) NOT NULL,
     ape VARCHAR(255) DEFAULT NULL,
     username VARCHAR(255) NOT NULL UNIQUE,
@@ -191,10 +191,10 @@ CREATE INDEX ix_usuario_id_role ON usuario (id_role);
 CREATE INDEX ix_usuario_nombre_completo ON usuario (name_user, ape);
 
 CREATE TABLE IF NOT EXISTS usuario_by_farmacia (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
 	name_farmacia TEXT DEFAULT NULL,
-    id_farmacia BIGINT,
-    id_usuario BIGINT NOT NULL,
+    id_farmacia INTEGER,
+    id_usuario INTEGER NOT NULL,
 
     CONSTRAINT fk_usuario_by_farmacia_id_farmacia FOREIGN KEY (id_farmacia) 
         REFERENCES farmacias (id) ON DELETE SET NULL ON UPDATE CASCADE,
